+++ Example/AndesUI/Home/Presenters/HomePresenter.swift
@@ -11,6 +11,7 @@ import UIKit
 protocol HomePresenter: NSObject {
     func presentButtons()
     func presentMessages()
+    func presentBadges()
     func presentWhatsNew()
 }
 
@@ -33,6 +34,11 @@ extension HomeViewPresenter: HomePresenter {
     func presentMessages() {
         router.routeToMessages()
     }
+
+    func presentBadges() {
+        router.routeToBadges()
+    }
+
     func presentWhatsNew() {
         router.routeToWhatsNew()
     }