+++ LibraryComponents/Classes/AndesBadge/Size/AndesBadgeSize.swift
@@ -1,0 +1,12 @@
+//
+//  AndesBadgeSize.swift
+//  AndesUI
+//
+
+/**
+ The AndesBadgeSize contains the differents sizes that a badge supports
+ */
+@objc public enum AndesBadgeSize: Int {
+    case small
+    case large
+}