+++ LibraryComponents/Classes/AndesBadge/Type/AndesBadgeTypeError.swift
@@ -1,0 +1,12 @@
+//
+//  AndesBadgeTypeError.swift
+//  AndesUI
+//
+
+import Foundation
+
+class AndesBadgeTypeError: AndesBadgeTypeProtocol {
+    var primaryColor: UIColor = UIColor.Andes.red500
+    var secondaryColor: UIColor = UIColor.Andes.red100
+    var icon: String? = AndesIcons.feedbackError16
+}