+++ LibraryComponents/Classes/AndesBadge/Type/AndesBadgeTypeProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  AndesBadgeTypeProtocol.swift
+//  AndesUI
+//
+
+import Foundation
+
+internal protocol AndesBadgeTypeProtocol {
+    var primaryColor: UIColor { get }
+    var secondaryColor: UIColor { get }
+    var icon: String? { get }
+}